TO TEST: constexpr'ed functions in tonc_math

+++ src/tte/tte_init_ase.cpp
@@ -78,7 +78,7 @@ void tte_init_ase(int bgnr, u16 bgcnt, u8 ase0, u32 clrs, u32 bupofs,
 				
 	// --- Bup font ---
 	void *dstD= &tile_mem[BFN_GET(bgcnt, BG_CBB)][2*ase0];
-	u32 dstS= font->charCount*font->cellSize;
+	u16 dstS= font->charCount*font->cellSize;
 
 	BUP bup= { dstS, font->bpp, 8, bupofs };
 	BitUnPack(font->data, dstD, &bup);
