+++ src/tte/tte_init_obj.cpp
@@ -75,7 +75,7 @@ void tte_init_obj(OBJ_ATTR *obj, u32 attr0, u32 attr1, u32 attr2,
 	//tc->cattr[3]= 0;
 
 	// --- Unpack font ---
-	u32 dstS= font->charCount*font->cellSize;
+	u16 dstS= font->charCount*font->cellSize;
 
 	BUP bup= { dstS, font->bpp, 4, bupofs };
 	BitUnPack(font->data, &tile_mem[4][tid], &bup);
